[analyzer] Move get_analyzer_checkers function to the config handler
