+++ analyzer/codechecker_analyzer/analyzers/clangsa/config_handler.py
@@ -11,12 +11,55 @@ from __future__ import print_function
 from __future__ import division
 from __future__ import absolute_import
 
+import re
+import subprocess
+
 from codechecker_common.logger import get_logger
 from .ctu_autodetection import CTUAutodetection
+
+from . import clang_options
+from . import version
 
 from .. import config_handler
 
 LOG = get_logger('analyzer.clangsa')
+
+
+def parse_checkers(clangsa_output):
+    """ Parse clang static analyzer checkers list output.
+
+    The given clangsa otput contains checker name and description pairs, which
+    however can extend over multiple lines. This is why we need stateful
+    parsing when iterating over lines.
+
+    Return a list of (checker name, description) tuples.
+    """
+    # Checker name and description in one line.
+    checker_entry_pattern = re.compile(
+        r'^\s\s(?P<checker_name>\S*)\s*(?P<description>.*)')
+
+    indented_pattern = re.compile(r'^\s\s\S')
+
+    checkers_list = []
+    checker_name = None
+    for line in clangsa_output.splitlines():
+        if line.startswith('CHECKERS:') or line == '':
+            continue
+        elif checker_name and not indented_pattern.match(line):
+            # Collect description for the checker name.
+            checkers_list.append((checker_name, line.strip()))
+            checker_name = None
+        elif re.match(r'^\s\s\S+$', line.rstrip()):
+            # Only checker name is in the line.
+            checker_name = line.strip()
+        else:
+            # Checker name and description is in one line.
+            match = checker_entry_pattern.match(line.rstrip())
+            if match:
+                current = match.groupdict()
+                checkers_list.append((current['checker_name'],
+                                      current['description']))
+    return checkers_list
 
 
 class ClangSAConfigHandler(config_handler.AnalyzerConfigHandler):
@@ -34,6 +77,42 @@ class ClangSAConfigHandler(config_handler.AnalyzerConfigHandler):
         self.enable_z3 = False
         self.enable_z3_refutation = False
         self.environ = environ
+
+    def get_analyzer_checkers(self, environ):
+        """
+        Return the list of the supported checkers.
+        """
+        analyzer_binary = self.analyzer_binary
+
+        try:
+            analyzer_version = subprocess.check_output(
+                [analyzer_binary, '--version'],
+                env=environ)
+
+        except subprocess.CalledProcessError as cerr:
+            LOG.error('Failed to get and parse clang version: %s',
+                      analyzer_binary)
+            LOG.error(cerr)
+            return []
+
+        version_parser = version.ClangVersionInfoParser()
+        version_info = version_parser.parse(analyzer_version)
+
+        command = [analyzer_binary, "-cc1"]
+
+        checkers_list_args = clang_options.get_analyzer_checkers_cmd(
+            version_info,
+            environ,
+            self.analyzer_plugins,
+            alpha=True)
+        command.extend(checkers_list_args)
+
+        try:
+            result = subprocess.check_output(command, env=environ,
+                                             universal_newlines=True)
+            return parse_checkers(result)
+        except (subprocess.CalledProcessError, OSError):
+            return []
 
     def add_checker_config(self, config):
         """
